[ refactor ] sharedPage hook 정리
