+++ src/utils/handleModal.js
@@ -1,5 +1,4 @@
 const handleModal = (setModal, modal) => {
-  console.log(`모달값이 ${modal}로 변경됐습니다.`);
   setModal(modal);
 };
